fix: pre-release hardening pass — all round-3 audit gaps
Security / correctness fixes:
- C1: WriteVarInt now panics for magnitudes > 55 bits (lib0 protocol
  ceiling); eliminates MinInt64 encode/decode roundtrip failure
- C2: YArray.Slice() guards start > end; prevents negative-capacity panic
- C3: persistence goroutine wrapped in defer/recover; StoreUpdate panic
  can no longer crash the WebSocket server
- H1: MaxConnections / MaxPeersPerRoom TOCTOU replaced with atomic
  counters; limits cannot be exceeded under concurrent upgrades
- H2: YText.Format() now sets OriginRight on ContentFormat items,
  matching Insert() and enabling correct YATA conflict resolution
- H3: YArray.Get() returns ContentType elements (nested shared types)
  instead of silently returning nil
- H4: EncodeUpdate(nil) includes removed (nil-State) clients, so
  peer-departure is broadcast correctly; docstring corrected
- H5: RleByteDecoder / UintOptRleDecoder / IntDiffOptRleDecoder guard
  uint64 repeat-count against MaxInt32 before int conversion
- M2: TransactContext returns ctx.Err() after Transact completes
- M3: checkJSONDepth returns false for unterminated strings
- M4: Awareness.Destroy() stops the auto-expiry goroutine; no more
  goroutine leak when the caller forgets the stop function
- M5: ReadAny tag-125 returns int64 (not int) preserving precision;
  NewContentAny normalises int → int64 for wire consistency

New features (Yjs API parity):
- A1: YText.ApplyDelta(txn, []Delta) — apply Quill-format deltas
- A2: YMap.ForEach(fn) — idiomatic iteration over live entries

Documentation:
- D1/D3: YText.Len() comment and test comment corrected to say
  "UTF-16 code units" instead of "runes" / "code points"
- D2: EncodeUpdate docstring matches implementation
- D4: Encoder / Decoder marked as not safe for concurrent use

Tests (10 new coverage areas):
- T1: Supplementary Unicode / emoji round-trips in YText
- T2: Nested shared types (YArray[YMap], YMap[YText]) integration tests
- T3: ErrTooManyClients and ErrStateTooLarge directed unit tests
- T4: TransactContext cancellation paths
- T5: UndoManager with YArray and YMap
- T7/T8: YArray.Slice(start>end) and Get(ContentType) regression tests
- T9: BenchmarkB4_Apply/EncodeV1/EncodeV2/DecodeV1 in benchmarks/
- T10: EncodeUpdate(nil) removal round-trip
- RLE overflow guard, VarInt range, int64 precision tests

+++ crdt/snapshot_test.go
@@ -210,7 +210,7 @@ func TestInteg_RestoreDocument_YArray(t *testing.T) {
 	// Restored should have only [1, 2, 3].
 	restored, err := RestoreDocument(doc, snap)
 	require.NoError(t, err)
-	assert.Equal(t, []any{1, 2, 3}, restored.GetArray("a").ToSlice())
+	assert.Equal(t, []any{int64(1), int64(2), int64(3)}, restored.GetArray("a").ToSlice())
 }
 
 func TestInteg_EncodeStateFromSnapshot(t *testing.T) {
